refactor: 테스트 코드 리팩토링 및 헬퍼 함수 추가
- 테스트 코드에서 중복된 부분을 헬퍼 함수로 분리하여 가독성 향상
- `SetupTestContext`, `SetupTestContextWithSession`, `CreateTestPosts` 등의 헬퍼 함수 추가
- 기존 테스트 케이스에서 헬퍼 함수를 활용하여 코드 간결화
- 불필요한 패키지 및 코드 제거로 테스트 파일 정리
