+++ internal/handler/handler_test/get_posts_test.go
@@ -1,89 +1,25 @@
-package handler_test
+package handler
 
 import (
 	"bumsiku/domain"
 	"bumsiku/internal/handler"
-	"bumsiku/internal/repository"
-	"context"
 	"encoding/json"
 	"net/http"
-	"net/http/httptest"
 	"testing"
-	"time"
 
-	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 )
-
-// Mock PostRepository
-type mockPostRepository struct {
-	posts     []domain.Post
-	nextToken *string
-	err       error
-}
-
-func (m *mockPostRepository) GetPosts(ctx context.Context, input *repository.GetPostsInput) (*repository.GetPostsOutput, error) {
-	if m.err != nil {
-		return nil, m.err
-	}
-	return &repository.GetPostsOutput{
-		Posts:     m.posts,
-		NextToken: m.nextToken,
-	}, nil
-}
-
-func (m *mockPostRepository) GetPostById(ctx context.Context, postId string) (*domain.Post, error) {
-	if m.err != nil {
-		return nil, m.err
-	}
-
-	for _, post := range m.posts {
-		if post.PostID == postId {
-			return &post, nil
-		}
-	}
-
-	return nil, nil
-}
-
-// 테스트용 데이터 생성
-func createTestPosts() []domain.Post {
-	now := time.Now()
-	return []domain.Post{
-		{
-			PostID:    "post1",
-			Title:     "첫 번째 게시글",
-			Content:   "내용 1",
-			Category:  "tech",
-			CreatedAt: now,
-			UpdatedAt: now,
-		},
-		{
-			PostID:    "post2",
-			Title:     "두 번째 게시글",
-			Content:   "내용 2",
-			Category:  "life",
-			CreatedAt: now.Add(time.Hour),
-			UpdatedAt: now.Add(time.Hour),
-		},
-	}
-}
 
 // [GIVEN] 정상적인 게시글 목록이 있는 경우
 // [WHEN] GetPosts 핸들러를 호출
 // [THEN] 상태코드 200과 게시글 목록 반환 확인
 func TestGetPosts_Success(t *testing.T) {
 	// Given
-	gin.SetMode(gin.TestMode)
-	mockPosts := createTestPosts()
+	mockPosts := CreateTestPosts()
 	mockRepo := &mockPostRepository{posts: mockPosts}
 
 	// When
-	w := httptest.NewRecorder()
-	c, _ := gin.CreateTestContext(w)
-	req := httptest.NewRequest("GET", "/posts", nil)
-	c.Request = req
-
+	c, w := SetupTestContext("GET", "/posts", "")
 	handler.GetPosts(mockRepo)(c)
 
 	// Then
@@ -102,16 +38,11 @@ func TestGetPosts_Success(t *testing.T) {
 // [THEN] 상태코드 200과 필터링된 게시글 목록 반환 확인
 func TestGetPosts_WithCategory(t *testing.T) {
 	// Given
-	gin.SetMode(gin.TestMode)
-	mockPosts := []domain.Post{createTestPosts()[0]} // tech 카테고리만
+	mockPosts := []domain.Post{CreateTestPosts()[0]} // tech 카테고리만
 	mockRepo := &mockPostRepository{posts: mockPosts}
 
 	// When
-	w := httptest.NewRecorder()
-	c, _ := gin.CreateTestContext(w)
-	req := httptest.NewRequest("GET", "/posts?category=tech", nil)
-	c.Request = req
-
+	c, w := SetupTestContext("GET", "/posts?category=tech", "")
 	handler.GetPosts(mockRepo)(c)
 
 	// Then
@@ -130,19 +61,14 @@ func TestGetPosts_WithCategory(t *testing.T) {
 // [THEN] 상태코드 200과 nextToken이 포함된 응답 반환 확인
 func TestGetPosts_WithPagination(t *testing.T) {
 	// Given
-	gin.SetMode(gin.TestMode)
 	nextToken := "next_page_token"
 	mockRepo := &mockPostRepository{
-		posts:     createTestPosts()[:1],
+		posts:     CreateTestPosts()[:1],
 		nextToken: &nextToken,
 	}
 
 	// When
-	w := httptest.NewRecorder()
-	c, _ := gin.CreateTestContext(w)
-	req := httptest.NewRequest("GET", "/posts?pageSize=1", nil)
-	c.Request = req
-
+	c, w := SetupTestContext("GET", "/posts?pageSize=1", "")
 	handler.GetPosts(mockRepo)(c)
 
 	// Then
@@ -160,22 +86,12 @@ func TestGetPosts_WithPagination(t *testing.T) {
 // [THEN] 상태코드 500과 에러 메시지 반환 확인
 func TestGetPosts_Error(t *testing.T) {
 	// Given
-	gin.SetMode(gin.TestMode)
 	mockRepo := &mockPostRepository{err: assert.AnError}
 
 	// When
-	w := httptest.NewRecorder()
-	c, _ := gin.CreateTestContext(w)
-	req := httptest.NewRequest("GET", "/posts", nil)
-	c.Request = req
-
+	c, w := SetupTestContext("GET", "/posts", "")
 	handler.GetPosts(mockRepo)(c)
 
 	// Then
-	assert.Equal(t, http.StatusInternalServerError, w.Code)
-
-	var response map[string]interface{}
-	err := json.Unmarshal(w.Body.Bytes(), &response)
-	assert.NoError(t, err)
-	assert.Equal(t, "게시글 목록 조회에 실패했습니다", response["error"])
+	AssertResponseJSON(t, w, http.StatusInternalServerError, "error", "게시글 목록 조회에 실패했습니다")
 }
